+++ include/system/Font.h
@@ -26,13 +26,13 @@ class Font {
     static unsigned int heightText(float scale, unsigned int maxWidth, std::string s);
 
     static void drawText(unsigned int x, unsigned int y, float scale,
-                         const unsigned char color[4], std::string s);
+                         glm::vec4 color, std::string s);
 
     static void drawTextWrapped(unsigned int x, unsigned int y, float scale,
-                                const unsigned char color[4], unsigned int maxWidth, std::string s);
+                                glm::vec4 color, unsigned int maxWidth, std::string s);
 
     static void drawTextCentered(unsigned int x, unsigned int y, float scale,
-                                 const unsigned char color[4], unsigned int width, std::string s);
+                                 glm::vec4 color, unsigned int width, std::string s);
 
     static void setShowFontBox(bool s) { showFontBox = s; }
     static bool getShowFontBox() { return showFontBox; }