+++ include/system/FontTTF.h
@@ -40,15 +40,15 @@ class FontTTF {
     static unsigned int heightText(float scale, unsigned int maxWidth, std::string s);
 
     static void drawText(unsigned int x, unsigned int y, float scale,
-                         const unsigned char color[4], std::string s);
+                         glm::vec4 color, std::string s);
     static void drawTextWrapped(unsigned int x, unsigned int y, float scale,
-                                const unsigned char color[4], unsigned int maxWidth, std::string s);
+                                glm::vec4 color, unsigned int maxWidth, std::string s);
 
   private:
     static int charIsMapped(int c);
     static int getQuad(int c, float* xpos, float* ypos, stbtt_aligned_quad* quad);
     static void drawTextInternal(unsigned int x, unsigned int y, float scale,
-                                 const unsigned char color[4], unsigned int maxWidth, std::string s,
+                                 glm::vec4 color, unsigned int maxWidth, std::string s,
                                  bool drawWrapped);
 
     static unsigned char* fontData;