+++ src/system/Font.cpp
@@ -62,7 +62,7 @@ unsigned int Font::heightText(float scale, unsigned int maxWidth, std::string s)
 }
 
 void Font::drawText(unsigned int x, unsigned int y, float scale,
-                    const unsigned char color[4], std::string s) {
+                    glm::vec4 color, std::string s) {
     if (stringEndsWith(fontName, ".pc")) {
         FontTRLE::drawText(x, y, scale, color, s);
     } else if (stringEndsWith(fontName, ".ttf")) {
@@ -79,7 +79,7 @@ void Font::drawText(unsigned int x, unsigned int y, float scale,
 }
 
 void Font::drawTextWrapped(unsigned int x, unsigned int y, float scale,
-                           const unsigned char color[4], unsigned int maxWidth, std::string s) {
+                           glm::vec4 color, unsigned int maxWidth, std::string s) {
     if (stringEndsWith(fontName, ".pc")) {
         FontTRLE::drawTextWrapped(x, y, scale, color, maxWidth, s);
     } else if (stringEndsWith(fontName, ".ttf")) {
@@ -98,7 +98,7 @@ void Font::drawTextWrapped(unsigned int x, unsigned int y, float scale,
 }
 
 void Font::drawTextCentered(unsigned int x, unsigned int y, float scale,
-                            const unsigned char color[4], unsigned int width, std::string s) {
+                            glm::vec4 color, unsigned int width, std::string s) {
     drawText(x + ((width / 2) - (widthText(scale, s) / 2)), y, scale, color, s);
 }